Migrate away query test from test helper infrastructure
